+++ src/utils/random-map-elem.js
@@ -1,3 +1,3 @@
-import {getRandomArrayElements} from "./random-array-elem.js";
+import {getRandomArrayElements} from "./random-array-elem";
 
 export const getRandomMapElements = (map, elementsAmount) => getRandomArrayElements(Array.from(map.keys()), elementsAmount);
